requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
redis>=5.0.0
python-dotenv>=1.0.0
langchain>=0.2
openai>=1.30.0
tiktoken>=0.6.0
pydantic>=2.0.0
